refactor: refactored navbar model to model folder
