+++ lib/features/interactive-book/models/navbar.dart
